Add files via upload
This is the E-prime file for this experiment. If you are not familiar with E-prime, it is almost like a composing slideshow, with each slide being a different type of object. Some slides are pure text, used for things such as instructions. Other slides contain code. Other slides have visuals.

This file includes a greeting and instructions, followed by the visual feedback system. There are three target forces--low, medium, and high--which can be defined by the researcher. The experiment is designed such that the participant alternates between the force-matching task at a specified target force and a rest period. The number and order of cycles can be defined in the list. Other parameters of interest, such as the max force correlating to the max height of the bar and the frequency of looping can be defined under the script titled Init0. 

The core component of this file, the visual feedback system, involves a loop that refreshes extremely quickly, each time updating the bar's height to reflect the most recent force data. Because the frequency of looping is so high, the visual feedback appears mostly smooth to the eye.